% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{between}
\alias{between}
\title{Shorthand for two numerical comparisons}
\usage{
between(x, gte, lte)
}
\arguments{
\item{x}{numeric value}

\item{gte}{lower bound}

\item{lte}{upper bound}
}
\value{
boolean
}
\description{
Shorthand for two numerical comparisons
}
