% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sided.R
\name{check_bounds_one_sided}
\alias{check_bounds_one_sided}
\title{Check Validity of One-Sided Bounds}
\usage{
check_bounds_one_sided(upper_bounds)
}
\arguments{
\item{upper_bounds}{Numeric vector where the ith component is the upper bound
for the ith order statistic.}
}
\value{
None
}
\description{
Given bounds for a one sided test, this checks that none of
the bounds fall outside of [0, 1].
}
