% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{check_bounds_two_sided}
\alias{check_bounds_two_sided}
\title{Check Validity of Two-Sided Bounds}
\usage{
check_bounds_two_sided(lower_bounds, upper_bounds)
}
\arguments{
\item{lower_bounds}{Numeric vector where the ith component is the lower bound
for the ith order statistic.}

\item{upper_bounds}{Numeric vector where the ith component is the lower bound
for the ith order statistic.}
}
\value{
None
}
\description{
Given bounds for a two sided test, this checks that none of
the bounds fall outside of [0, 1] and that all upper bounds
are greater than the corresponding lower bounds.
This also ensures the the length of the bounds are the same.
This not meant to be called by the user.
}
