% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_dparams}
\alias{estimate_dparams}
\title{Estimate parameters of distribution given set of observations}
\usage{
estimate_dparams(distribution, obs)
}
\arguments{
\item{distribution}{The quantile  or probability function for the specified distribution given as a string.
For example \code{"qnorm"} or \code{"pnorm"}.}

\item{obs}{The observed data.}
}
\value{
list of estimated parameters
}
\description{
For the normal distribution, we estimate the mean as the median and the standard deviation as \eqn{Sn} from the paper by Rousseeuw and Croux 1993
  "Alternatives to the Median Absolute Deviation". For all other distributions besides normal,
  the code uses MLE to estimate the parameters.
}
