% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_params_from_data}
\alias{estimate_params_from_data}
\title{Estimate Parameters from Data}
\usage{
estimate_params_from_data(distr_name, obs)
}
\arguments{
\item{distr_name}{\code{MASS} name of distribution}

\item{obs}{observation vector}
}
\value{
list of distribution parameters
}
\description{
For select distributions, parameters are estimated from data. Generally,
the MLEs are used. However, for the normal distribution we use robust
estimators.
}
