% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{get_asymptotic_approx_corrected_alpha}
\alias{get_asymptotic_approx_corrected_alpha}
\title{Calculates Approximate Local Level}
\usage{
get_asymptotic_approx_corrected_alpha(n, alpha)
}
\arguments{
\item{n}{Number of tests to do}

\item{alpha}{Global type I error rate \eqn{\alpha} of the tests}
}
\value{
Approximate local level
}
\description{
This function uses the approximation from Gontscharuk & Finner's Asymptotics of
goodness-of-fit tests based on minimum p-value statistics (2017) to approximate
local levels for finite sample size. We use these authors constants for \eqn{\alpha} = .1, and .05,
and for \eqn{\alpha} = .01 we use a slightly different approximation.
}
