% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_best_available_prob_pts_method}
\alias{get_best_available_prob_pts_method}
\title{Get Best Available Method for Probability Points}
\usage{
get_best_available_prob_pts_method(dist_name)
}
\arguments{
\item{dist_name}{character name of distribution}
}
\value{
character name of best expected points method
}
\description{
Determines name of best method for obtaining expected points for a qq or pp
plot.
}
