% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sided.R
\name{get_bounds_one_sided}
\alias{get_bounds_one_sided}
\title{Calculates Local Bounds For One-Sided Test}
\usage{
get_bounds_one_sided(alpha, n, tol = 1e-06, max_it = 100)
}
\arguments{
\item{alpha}{Desired Type I Error rate of the complete test.}

\item{n}{Size of the dataset.}

\item{tol}{(Optional) Relative tolerance of the \eqn{\alpha} level of the
simultaneous test. Defaults to 1e-8.}

\item{max_it}{(Optional) Maximum number of iterations of Binary Search Algorithm
used to find the bounds. Defaults to 100 which should be much larger than necessary
for a reasonable tolerance.}
}
\value{
A list with components
\itemize{
  \item bound - Numeric vector containing the bounds of the test of each order statistic.
  \item x - Expectation of each order statistic. This is the x-axis for the bounds if used in a qq-plot.
  \item local_level - Type I Error rate of each individual test on the order statistic. It is equal for all tests.
}
}
\description{
Given the size of the dataset and the desired Type I Error rate \eqn{\alpha},
this calculates one-sided testing bounds for each order statistic.
These bounds are created using "equal local levels," which means that
each bound corresponds to a test that has a Type I Error rate of \eqn{\eta}
on its corresponding order statistic. Note that this code may be very slow for
values of n.
}
\examples{
get_bounds_one_sided(alpha = .05, n = 10, tol = 1e-6, max_it = 50)


}
