% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{get_bounds_two_sided}
\alias{get_bounds_two_sided}
\title{Calculates Rejection Region of Two-Sided Equal Local Levels Test.}
\usage{
get_bounds_two_sided(alpha, n, tol = 1e-08, max_it = 100,
  method = c("best_available", "approximate", "search"))
}
\arguments{
\item{alpha}{Desired global significance level of the test.}

\item{n}{Size of the dataset.}

\item{tol}{(Optional) Relative tolerance of the \eqn{\alpha} level of the
simultaneous test. Defaults to 1e-8.}

\item{max_it}{(Optional) Maximum number of iterations of Binary Search Algorithm
used to find the bounds. Defaults to 100 which should be much larger than necessary
for a reasonable tolerance.}

\item{method}{(Optional) Parameter indicating if the calculation should be done using a highly
accurate approximation, "approximate", or if the calculations should be done using an exact
binary search calculation, "search". The default is "best_available" (recommended), which uses the exact
search when the approximation isn't highly accurate and the search method isn't prohibitively slow. Of note, 
the approximate method is only usable for alpha values of .1, .05, and .01 (within 10 ^ (-5)).}
}
\value{
A list with components
\itemize{
  \item lower_bound - Numeric vector containing the lower bounds of the test of each order statistic.
  \item upper_bound - Numeric vector containing the upper bounds of the test of each order statistic.
  \item x - Expectation of each order statistic. This is the x-axis for the bounds if used in a qq-plot.
  \item local_level - Type I Error rate of each individual test on the order statistic. It is equal for all tests.
}
}
\description{
The context is that n i.i.d. observations are assumed to be drawn
from some distribution on the unit interval with c.d.f. F(x), and it is 
desired to test the null hypothesis that F(x) = x for all x in (0,1), 
referred to as the "global null hypothesis," against a two-sided alternative.
An "equal local levels" test is used, in which each of the n order statistics is
tested for significant deviation from its null distribution by a 2-sided test 
with significance level eta.  The global null hypothesis is rejected if at 
least one of the order statistic tests is rejected at level eta, where eta is 
chosen so that the significance level of the global test is alpha.  
Given the size of the dataset n and the desired global significance level alpha, 
this function calculates the local level eta and the acceptance/rejection regions for the test.
There are a set of n intervals, one for each order statistic.  
If at least one order statistic falls outside the corresponding interval, 
the global test is rejected.
}
\examples{
get_bounds_two_sided(alpha = .05, n = 100, tol = 1e-6, max_it = 50)


}
