% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extended_quantile}
\alias{get_extended_quantile}
\title{Get Quantile for First and Last Point of QQ or PP Plot}
\usage{
get_extended_quantile(exp_pts_method, n)
}
\arguments{
\item{exp_pts_method}{method used to derive expected points}

\item{n}{sample size}
}
\value{
list with low and high point
}
\description{
Get Quantile for First and Last Point of QQ or PP Plot
}
