% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mass_name_from_distr}
\alias{get_mass_name_from_distr}
\title{Convert R Distribution Function to MASS Distribution Name}
\usage{
get_mass_name_from_distr(distr, band_type)
}
\arguments{
\item{distr}{R distribution function (e.g. qnorm or pnorm)}

\item{band_type}{one of "qq" (for quantile functions) or "pp" (
for probability functions).}
}
\value{
string of MASS distribution name
}
\description{
Convert R Distribution Function to MASS Distribution Name
}
