% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{get_qq_bounds}
\alias{get_qq_bounds}
\title{Create QQ or PP Plot Testing Bounds}
\usage{
get_qq_bounds(
  n,
  obs,
  alpha,
  distribution,
  dparams = list(),
  ell_params = list(),
  method = c("ell", "ks", "pointwise"),
  band_type = c("qq", "pp")
)
}
\arguments{
\item{n}{(Optional) number of observations. If \code{obs} is not provided,
this must be specified. Providing \code{n} is simpler when all parameters
of \code{distribution} are known and specified via \code{dparams}
(or when using all default parameter choices of \code{distribution} is desired).}

\item{obs}{(Optional) numeric vector of observations. If \code{n} is not
provided, this must be specified. Supplying this argument is only recommended
when estimating parameters of \code{distribution} from data is desired.}

\item{alpha}{desired significance level of the testing bands. If \code{method}
is set to \code{"ell"} or \code{"ks"}, then this is the global significance
of the testing bands. If \code{method} is set to \code{"pointwise"}, then
the bands are equivalent to simply conducting a test on each order statistic
with Type I error \code{alpha}. Pointwise bands will generally have much
larger global Type I error than \code{alpha}.}

\item{distribution}{either a quantile function (e.g. \code{qnorm}) for
constructing bands on a QQ-plot or a probability function (e.g. \code{pnorm})
for constructing bands on a PP-plot. \code{band_type} must be set
appropriately.}

\item{dparams}{(Optional)  List of additional parameters for the \code{distribution} function
(e.g. df=1). Note that if any parameters of the distribution are specified, parameter estimation will not be performed
on the unspecified parameters, and instead they will take on the default values set by the distribution function.
For the uniform distribution, parameter estimation is not performed, and
the default parameters are max = 1 and min = 0.
For other distributions parameters will be estimated if not provided.
For the normal distribution, we estimate the mean as the median and the standard deviation as \eqn{Sn} from the paper by Rousseeuw and Croux 1993
"Alternatives to the Median Absolute Deviation". For all other distributions besides uniform and normal,
the code uses MLE to estimate the parameters. Note that estimation is not implemented for custom distributions, so all
parameters of the distribution must be provided by the user.}

\item{ell_params}{(Optional) list of optional parameters for \code{get_bounds_two_sided}
(i.e. \code{tol}, \code{max_it}, \code{method}). Only used if \code{method}
is set to \code{"ell"}}

\item{method}{(Optional) method for creating testing bands. The default,
\code{"ell"} uses the equal local levels method
(see \code{get_bounds_two_sided} for more information). \code{"ks"} uses
the Kolmogorov-Smirnov test. \code{"pointwise"} uses pointwise bands (see
documentation for parameter \code{alpha} for more information). \code{"ell"}
is recommended.}

\item{band_type}{(Optional) string specifying if the bands should be created
for a Quantile-Quantile (QQ) plot or a Probability-Probability (PP) plot.}
}
\value{

}
\description{
Create QQ or PP Plot Testing Bounds
}
