% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_qq_distribution_from_pp_distribution}
\alias{get_qq_distribution_from_pp_distribution}
\title{Get Quantile Distribution from Probability Distribution}
\usage{
get_qq_distribution_from_pp_distribution(dname)
}
\arguments{
\item{dname}{probability distribution (e.g. \code{pnorm})}
}
\value{
quantile distribution (e.g. \code{qnorm}).
}
\description{
Get Quantile Distribution from Probability Distribution
}
