% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_qq_band.R, R/stat_qq_band.R
\name{geom_qq_band}
\alias{geom_qq_band}
\alias{stat_qq_band}
\title{Quantile-quantile confidence bands}
\usage{
geom_qq_band(mapping = NULL, data = NULL, stat = "qq_band",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, distribution = "norm", dparams = list(),
  difference = FALSE, bandType = "pointwise", conf = 0.95, ...)

stat_qq_band(mapping = NULL, data = NULL, geom = "qq_band",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, distribution = "norm", dparams = list(),
  difference = FALSE, bandType = "pointwise", bounds.params = NULL,
  conf = 0.95, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{statistic to use to calculate confidence bands. Should be
`qq_band`.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{distribution}{Character. Theoretical probability distribution function
to use. Do not provide the full distribution function name (e.g.,
\code{"dnorm"}). Instead, just provide its shortened name (e.g.,
\code{"norm"}). If you wish to provide a custom distribution, you may do so
by first creating the density, quantile, and random functions following the
standard nomenclature from the \code{stats} package (i.e., for
\code{"custom"}, create the \code{dcustom}, \code{pcustom},
\code{qcustom}, and \code{rcustom} functions).}

\item{dparams}{List of additional parameters passed on to the previously
chosen \code{distribution} function. If an empty list is provided (default)
then the distributional parameters are estimated via MLE. MLE for custom
distributions is currently not supported, so you must provide the
appropriate \code{dparams} in that case.}

\item{difference}{Logical. Should the plot objects be differenceed? If \code{TRUE},
the objects will be differenceed according to the reference Q-Q line. This
procedure was described by Thode (2002), and may help reducing visual bias
caused by the orthogonal distances from Q-Q points to the reference line.}

\item{bandType}{Character. Either \code{"pointwise"}, \code{"boot"}, \code{"ks"} or
\code{"ts"}. \code{"pointwise"} constructs pointwise confidence bands based
on Normal confidence intervals. \code{"boot"} creates pointwise confidence
bands based on a parametric bootstrap; parameters are estimated with MLEs.
\code{"ks"} constructs simultaneous confidence bands based on the Kolmogorov-Smirnov
test. Finally, \code{"ts"} constructs tail-sensitive confidence bands, as
described by Aldor-Noiman et al. (2013) (also, see 'Note' for
limitations).}

\item{conf}{Numerical. Confidence level of the bands.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{geom}{The geometric object to use display the data}

\item{bounds.params}{List of optional parameters for get_bounds_two_sided
(i.e. tol, max_it, method).}
}
\description{
Draws quantile-quantile confidence bands, with an additional difference option.
}
