% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{get_bounds_two_sided}
\alias{get_bounds_two_sided}
\title{Calculates Local Bounds For Two-Sided Test}
\usage{
get_bounds_two_sided(alpha, n, tol = 1e-08, max_it = 100,
  method = "approximate")
}
\arguments{
\item{alpha}{Desired Type I Error rate of the complete test.}

\item{n}{Size of the dataset.}

\item{tol}{(Optional) Relative tolerance of the \eqn{\alpha} level of the
simultaneous test. Defaults to 1e-8.}

\item{max_it}{(Optional) Maximum number of iterations of Binary Search Algorithm
used to find the bounds. Defaults to 100 which should be much larger than necessary
for a reasonable tolerance.}

\item{method}{(Optional) Parameter indicating if the calculation should be done using a highly
accurate approximation, "approximate", or if the calculations should be done using an exact
binary search calculation, "search". The approximate method is only usable for alpha values
of .1, .05, and .01.}
}
\value{
A list with components
\itemize{
  \item lower_bound - Numeric vector containing the lower bounds of the test of each order statistic.
  \item upper_bound - Numeric vector containing the upper bounds of the test of each order statistic.
  \item x - Expectation of each order statistic. This is the x-axis for the bounds if used in a qq-plot.
  \item local_level - Type I Error rate of each individual test on the order statistic. It is equal for all tests.
}
}
\description{
Given the size of the dataset and the desired Type I Error rate \eqn{\alpha},
this calculates two-sided testing bounds for each order statistic.
These bounds are created using "equal local levels," which means that
each bound pair corresponds to a test that has a Type I Error rate of \eqn{\eta}
on its corresponding order statistic.
}
\examples{
get_level_from_bounds_twosided(alpha = .05, n = 100, tol = 1e-6, max_it = 50)


}
