% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_sided.R
\name{get_level_from_bounds_one_sided}
\alias{get_level_from_bounds_one_sided}
\title{Calculates Type I Error Rate From One-Sided Bounds}
\usage{
get_level_from_bounds_one_sided(bounds)
}
\arguments{
\item{bounds}{Numeric vector where the ith component is the lower bound
for the ith order statistic. The values must be in ascending order.}
}
\value{
Type I Error Rate \eqn{\alpha}
}
\description{
Given bounds for a one sided test on uniform order statistics, this computes
the Type I Error Rate \eqn{\alpha} using a binary search. The function also
allows for an approximate computation for speed with some relative tolerance.
}
\examples{
get_level_from_bounds_one_sided(bounds = c(.1, .5, .8))

}
