% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{get_level_from_bounds_two_sided}
\alias{get_level_from_bounds_two_sided}
\title{Calculates Global Significance Level From Simultaneous Two-Sided Bounds for Rejection Region}
\usage{
get_level_from_bounds_two_sided(lower_bounds, upper_bounds)
}
\arguments{
\item{lower_bounds}{Numeric vector where the ith component is the lower bound for the acceptance interval
for the ith order statistic. The values must be in ascending order.}

\item{upper_bounds}{Numeric vector of the same length as \code{lower_bounds} where the ith component is the upper bound
for the acceptance interval for the ith order statistic. The values must be in ascending order and
the ith component must be larger than the ith component of the \code{lower_bounds} vector.}
}
\value{
Global Significance Level \eqn{\alpha}
}
\description{
For a test of uniformity of i.i.d. observations on the unit interval, this function will determine the significance
level as a function of the rejection region. Suppose \eqn{n} observations are drawn i.i.d. from some CDF F(x) on the unit interval,
and it is desired to test the null hypothesis that F(x) = x for all x in (0, 1) against a two-sided alternative. 
Suppose the acceptance region for the test is described by a set of intervals, one for each order statistic.
Given the bounds for these intervals, this function calculates the significance level of the test where the
null hypothesis is rejected if at least one of the order statistics is outside its corresponding interval.
}
\examples{
# For X1, X2 iid unif(0,1), calculate 1 - P(.1 < min(X1, X2) < .6 and .5 < max(X1, X2) < .9)
get_level_from_bounds_two_sided(lower_bounds = c(.1, .5), upper_bounds = c(.6, .9))

# Finds the global significance level corresponding to the local level eta
n <- 100
eta <- .05
lb <- qbeta(eta / 2, c(1:n), c(n:1))
ub <- qbeta(1 - eta / 2, c(1:n), c(n:1))
get_level_from_bounds_two_sided(lower_bounds = lb, upper_bounds = ub)

}
