% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{get_level_from_bounds_two_sided}
\alias{get_level_from_bounds_two_sided}
\title{Calculates Type I Error Rate From Two-Sided Bounds}
\usage{
get_level_from_bounds_two_sided(lower_bounds, upper_bounds)
}
\arguments{
\item{lower_bounds}{Numeric vector where the ith component is the lower bound
for the ith order statistic. The values must be in ascending order.}

\item{upper_bounds}{Numeric vector where the ith component is the lower bound
for the ith order statistic. The values must be in ascending order and
the ith component must be larger than the ith component of the lower bounds.}
}
\value{
Type I Error Rate \eqn{\alpha}
}
\description{
Given bounds for a two sided test on uniform order statistics, this computes
the Type I Error Rate \eqn{\alpha} using a binary search.
}
\examples{
get_level_from_bounds_two_sided(lower_bounds = c(.1, .5), upper_bounds = c(.6, .9))

}
