% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sided.R
\name{get_qq_bounds}
\alias{get_qq_bounds}
\title{Create QQ or PP Plot Testing Bounds}
\usage{
get_qq_bounds(
  n,
  obs,
  alpha = 0.05,
  distribution = qnorm,
  dparams = list(),
  ell_params = list(),
  method = c("ell", "ks", "pointwise"),
  band_type = c("qq", "pp")
)
}
\arguments{
\item{n, obs}{either a number of observations (specified by setting \code{n}),
or a numeric vector of observations (specified by setting \code{obs}). One
argument must be specified. In principle, if all parameters \code{distribution}
are known, then testing bands only depend on the number of observations \code{n}.
Thus, providing \code{n} is simpler when all parameters
of \code{distribution} are known and specified via \code{dparams}
(or when using all default parameter choices of \code{distribution} is desired).
If estimating parameters from the data is preferred, \code{obs} should be
specified and estimation will take place.}

\item{alpha}{(optional) desired significance level of the testing bands. If \code{method}
is set to \code{"ell"} or \code{"ks"}, then this is the global significance
of the testing bands. If \code{method} is set to \code{"pointwise"}, then
the bands are equivalent to simply conducting a test on each order statistic
with Type I error \code{alpha}. Pointwise bands will generally have much
larger global Type I error than \code{alpha}. Defaults to \code{.05}.}

\item{distribution}{(optional) either a quantile function (e.g. \code{qnorm}) for
constructing bands on a QQ-plot or a probability function (e.g. \code{pnorm})
for constructing bands on a PP-plot. \code{band_type} must be set
appropriately. Defaults to \code{qnorm} for a normal distribution.}

\item{dparams}{(optional)  List of additional parameters for the \code{distribution} function
(e.g. df=1). Note that if any parameters of the distribution are specified, parameter estimation will not be performed
on the unspecified parameters, and instead they will take on the default values set by the distribution function.
For the uniform distribution, parameter estimation is not performed, and
the default parameters are max = 1 and min = 0.
For other distributions parameters will be estimated if not provided.
For the normal distribution, we estimate the mean as the median and the standard deviation as \eqn{Sn} from the paper by Rousseeuw and Croux 1993
"Alternatives to the Median Absolute Deviation". For all other distributions besides uniform and normal,
the code uses MLE to estimate the parameters. Note that estimation is not implemented for custom distributions, so all
parameters of the distribution must be provided by the user.}

\item{ell_params}{(optional) list of optional parameters for \code{get_bounds_two_sided}
(i.e. \code{tol}, \code{max_it}, \code{method}). Only used if \code{method}
is set to \code{"ell"}}

\item{method}{(optional) method for creating testing bands. The default,
\code{"ell"} uses the equal local levels method
(see \code{get_bounds_two_sided} for more information). \code{"ks"} uses
the Kolmogorov-Smirnov test. \code{"pointwise"} uses pointwise bands (see
documentation for parameter \code{alpha} for more information). \code{"ell"}
is recommended.}

\item{band_type}{(optional) string specifying if the bands should be created
for a Quantile-Quantile (QQ) plot or a Probability-Probability (PP) plot.}
}
\value{
A list with components
\itemize{
  \item lower_bound - Numeric vector of length \code{n} containing the lower bounds
  for the acceptance regions of the test corresponding to each order statistic.
  \item upper_bound - Numeric vector of length \code{n} containing the upper bounds
  for the acceptance regions of the test corresponding to each order statistic.
  \item expected_value - Numeric vector of length \code{n} containing the expectation of each order statistic.
  These are the x-coordinates for the bounds if used in a qq-plot.
}
}
\description{
Flexible interface for creating testing bounds for Quantile-Quantile (QQ) or
Probability-Probability (PP) plots.
}
\examples{

# Get ell level .05 QQ testing bounds for normal(0, 1) distribution with 100 observations
bounds <- get_qq_bounds(n = 100)

# Get ell level .05 QQ testing bounds for normal distribution with unknown parameters
obs <- rnorm(100)
bounds <- get_qq_bounds(obs = obs)

# Get ell level .05 PP testing bounds for t(2) distribution with 100 observations
bounds <- get_qq_bounds(
  n = 100, distribution = pt, dparams = list(df = 1), band_type = 'pp'
)

}
