% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppplot.R
\name{pp_conf_plot}
\alias{pp_conf_plot}
\title{PP Plot with Simultaneous and Pointwise Testing Bounds.}
\usage{
pp_conf_plot(
  obs,
  distribution = pnorm,
  method = c("ell", "ks"),
  alpha = 0.05,
  difference = FALSE,
  log10 = FALSE,
  add = FALSE,
  dparams = list(),
  bounds_params = list(),
  line_params = list(),
  plot_pointwise = FALSE,
  pointwise_lines_params = list(),
  points_params = list(),
  polygon_params = list(border = NA, col = "gray"),
  ...
)
}
\arguments{
\item{obs}{The observed data.}

\item{distribution}{The probability function for the specified distribution. Defaults to qnorm.
Custom distributions are allowed so long as all parameters are supplied in dparams.}

\item{method}{Method for simultaneous testing bands. Must be either "ell" (equal local levels test), which applies a level \eqn{\eta} pointwise
test to each order statistic such that the Type I error of the global test is \eqn{\alpha}, or "ks" to apply a 
Kolmogorov-Smirnov test. For \eqn{\alpha} = .01, .05, and .1, "ell" is recommended.}

\item{alpha}{Type I error of global test of whether the data come from the reference distribution.}

\item{difference}{Whether to plot the difference between the observed and
expected values on the vertical axis.}

\item{log10}{Whether to plot axes on -log10 scale (e.g. to see small p-values). Can only be used for strictly
positive distributions.}

\item{add}{Whether to add points to an existing plot.}

\item{dparams}{List of additional parameters for the quantile function of the distribution
(e.g. df=1). Note that if any parameters of the distribution are specified, all other unspecified parameters of the distribution
will take on their default values. For the uniform distribution, parameter estimation is not supported, and
the default parameters are max = 1 and min = 0.
For other distributions parameters will be estimated if not provided and an appropriate estimation procedure exists.
For the normal distribution, we estimate the mean as the median and the standard deviation as \eqn{Sn} from the paper by Rousseeuw and Croux 1993
"Alternatives to the Median Absolute Deviation". For all other distributions besides uniform and normal,
the code uses MLE to estimate the parameters. Note that estimation is not implemented for custom distributions, so all
parameters of the distribution must be provided by the user.}

\item{bounds_params}{List of optional parameters for get_bounds_two_sided
(i.e. tol, max_it, method).}

\item{line_params}{Parameters passed to the line function to modify the line that indicates a perfect fit of the
reference distribution.}

\item{plot_pointwise}{Boolean indiciating if pointwise bounds should be added to the plot}

\item{pointwise_lines_params}{Parameters passed to the \code{lines} function that modifies pointwise bounds if plot_pointwise is
set to TRUE.}

\item{points_params}{Parameters to be passed to the \code{points} function to plot the data.}

\item{polygon_params}{Parmeters to be passed to the polygon function to construct simultaneous confidence region.
By default \code{border} is set to NA and \code{col} is set to grey.}

\item{...}{Additional parameters passed to the plot function.}
}
\description{
Create a pp-plot with with a shaded simultaneous acceptance region and,
optionally, lines for a point-wise region. The observed values are plotted
against their expected values had they come from the specified distribution.
}
\details{
If any of the points of the pp-plot fall outside the simultaneous acceptance region for the selected
level alpha test, that means that we can reject the null hypothesis that the data are i.i.d. draws from the
specified distribution. If \code{difference} is set to TRUE, the vertical axis plots the 
observed quantile minus expected quantile. If pointwise bands are used, then on average, alpha * n of the points will fall outside
the bounds under the null hypothesis, so the chance that the pp-plot has any points falling outside of the pointwise bounds
is typically much higher than alpha under the null hypothesis. For this reason, a simultaneous region is preferred.
}
\examples{
set.seed(0)
smp <- rnorm(100)

# Plot PP plot against normal distribution with mean and variance estimated
pp_conf_plot(
  obs=smp, 
  distribution = pnorm
)

# Make same plot on -log10 scale to highlight small p-values
pp_conf_plot(
  obs=smp, 
  distribution = pnorm,
  log10 = TRUE
)

# Make same plot with difference between observed and expected values on the y-axis 
pp_conf_plot(
  obs=smp, 
  distribution = pnorm,
  difference = TRUE
)

# Make same plot with samples plotted as a blue line and expected value line plotted as a red line
pp_conf_plot(
  obs=smp, 
  distribution = pnorm,
  plot_pointwise = TRUE,
  method = "ell",
  points_params = list(col="blue", type="l"),
  line_params = list(col="red")
)

}
