% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qq_conf_plot}
\alias{qq_conf_plot}
\title{QQ Plot with Simultaneous and Pointwise Testing Bounds.}
\usage{
qq_conf_plot(obs, dist = qunif, method = c("ell", "ks"),
  alpha = 0.05, difference = FALSE, log10 = FALSE,
  shade.col = "gray", add = FALSE, dparams = NULL,
  bounds.params = NULL, pw.lty = 0, pw.col = "black", ...)
}
\arguments{
\item{obs}{The observed data.}

\item{dist}{The quantile function for the specified distribution. Defaults to qnorm.
Custom distributions are allowed so long as all parameters are supplied in dparams.}

\item{method}{Method for simultaneous testing bands. Must be either "ell", which applies a level \eqn{\eta} pointwise
test to each order statistic such that the Type I error of the global test is \eqn{\alpha}, or "ks" to apply a 
Kolmogorov-Smirnov test. For \eqn{\alpha} = .01, .05, and .1, "ell" is recommended.}

\item{alpha}{Type I error of global test of if the data comes from the reference distribution.}

\item{difference}{Whether to plot the difference between the observed and
expected values on the vertical axis.}

\item{log10}{Whether to plot axes on -log10 scale (e.g. to see small p-values).}

\item{shade.col}{What color to use for the simultaneous acceptance region.}

\item{add}{Whether to add points to an existing plot.}

\item{dparams}{List of additional parameters for the quantile function of the distribution
(e.g. df=1). Will be estimated if not provided. For the uniform distribution, U(0, 1) is assumed unless specified.
For the normal distribution, we estimate the mean
as the median and the standard deviation as \eqn{S_{n}} from the paper by Rousseeuw and Croux 1993
"Alternatives to the Median Absolute Deviation". For all other distributions,
the code uses MLE to estimate the parameters. Note that estimation is not implemented for custom distributions.}

\item{bounds.params}{List of optional parameters for get_bounds_two_sided
(i.e. tol, max_it, method).}

\item{pw.lty}{Line type for the pointwise error bounds. Set to non-zero for a line.}

\item{pw.col}{Color for the pointwise bounds line.}

\item{...}{Additional parameters for the plot.}
}
\description{
Create a qq-plot with with a shaded simultaneous acceptance region and,
optionally, lines for a point-wise region. The observed values are plotted
against their expected values had they come from the specified distribution.
}
\details{
For $N$ independent tests that follow a specified distribution, the global
confidence interval of size $C$ is the region such that the probability of
all the points being in the region is $C$. The pointwise bounds are determined
by taking the marginal probability that an individual point is within the bounds
is $C$. If 'difference' is set to TRUE, the vertical axis plots the 
observed quantile minus expected quantile. Set pw.lty to a non-zero line type to plot
the pointwise bounds.
}
\examples{
x <- rchisq(1000, 1)
qq_conf_plot(x, qchisq, dparams=list(df=1), pw.lty=3) # Plots x against a 1-df chisquare

y <- runif(893)
qq_conf_plot(y, difference = TRUE, log10 = TRUE, bounds.params = list(method = "search"), pch=3)
}
