% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_qq_band.R, R/stat_qq_band.R
\name{geom_qq_band}
\alias{geom_qq_band}
\alias{stat_qq_band}
\title{Quantile-quantile testing bands}
\usage{
geom_qq_band(mapping = NULL, data = NULL, stat = "qq_band",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, distribution = "norm", dparams = list(),
  difference = FALSE, method = "ell", alpha = 0.05, ...)

stat_qq_band(mapping = NULL, data = NULL, geom = "qq_band",
  position = "identity", na.rm = TRUE, show.legend = NA,
  inherit.aes = TRUE, distribution = "norm", dparams = list(),
  difference = FALSE, method = "ell", bounds_params = NULL,
  alpha = 0.05, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame.}, and
will be used as the layer data.}

\item{stat}{statistic to use to calculate confidence bands. Should be
`qq_band`.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{distribution}{The quantile function for the specified distribution. Defaults to qnorm.
Custom distributions are allowed so long as all parameters are supplied in dparams.}

\item{dparams}{List of additional parameters for the quantile function of the distribution
(e.g. df=1). Will be estimated if not provided and an appropriate estimation procedure exists.
For the normal distribution, we estimate the mean as the median and the standard deviation as \eqn{Sn} from the paper by Rousseeuw and Croux 1993
"Alternatives to the Median Absolute Deviation". For all other distributions,
the code uses MLE to estimate the parameters. Note that estimation is not implemented for custom distributions, so all
parameters of the distribution must be provided by the user.}

\item{difference}{Whether to plot the difference between the observed and
expected values on the vertical axis.}

\item{method}{Method for simultaneous testing bands. Must be either "ell", which applies a level \eqn{\eta} pointwise
test to each order statistic such that the Type I error of the global test is \eqn{\alpha}, or "ks" to apply a 
Kolmogorov-Smirnov test. For \eqn{\alpha} = .01, .05, and .1, "ell" is recommended.}

\item{alpha}{Type I error of global test of if the data comes from the reference distribution.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{geom}{The geometric object to use display the data}

\item{bounds_params}{List of optional parameters for get_bounds_two_sided
(i.e. tol, max_it, method).}
}
\description{
Draws quantile-quantile confidence bands, with an additional difference option.
}
\details{
If any of the points of the qq-plot fall outside the simultaneous acceptance region for the selected
level alpha test, that means that we can reject the null hypothesis that the data are i.i.d. draws from the
specified distribution. If 'difference' is set to TRUE, the vertical axis plots the 
observed quantile minus expected quantile. Set pw.lty to a non-zero line type to plot
the pointwise bounds. If pointwise bands are used, then on average, alpha * n of the points will fall outside
the bounds under the null hypothesis, so the chance that the qq-plot has any points falling outside of the pointwise bounds
is typically much higher than alpha under the null hypothesis. For this reason, a simultaneous region is preferred.
}
